+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 09 November 2024 - v1.0.2 ##
+* dynamic data on footer
+* show whatsapp number
+
 ## 07 November 2024 - v1.0.1 ##  
 * add read more link